# postman/postman-collection-openapi.json
{
    "openapi": "3.0.0",
    "paths": {
        "/auth/register": {
            "post": {
                "operationId": "register",
                "parameters": [],
                "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterDto" } } } },
                "responses": { "201": { "description": "" } },
                "tags": ["Auth"]
            }
        },
        "/auth/login": {
            "post": {
                "operationId": "login",
                "parameters": [],
                "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginDto" } } } },
                "responses": { "200": { "description": "" } },
                "tags": ["Auth"]
            }
        },
        "/auth/refresh": {
            "post": {
                "operationId": "refreshToken",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenDto" } } }
                },
                "responses": { "200": { "description": "" } },
                "tags": ["Auth"]
            }
        },
...
    },
    "info": {
        "title": "eCommerce API",
        "description": "Comprehensive eCommerce API with advanced querying capabilities including pagination, filtering, sorting, searching, and joins",
        "version": "1.0.0",
        "contact": { "name": "API Support", "url": "https://github.com/yourusername/ecommerce-api", "email": "support@yourcompany.com" },
        "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
    },
...
    "servers": [{ "url": "http://localhost:3151", "description": "Development server" }],
    "components": {
        "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
        "schemas": {
            "RegisterDto": { "type": "object", "properties": {} },
            "LoginDto": { "type": "object", "properties": {} },
            "RefreshTokenDto": { "type": "object", "properties": {} },
...
        }
    }
}